Major refactoring to healthcheck types
This will allow a consumer of the library to have their own continuous
healthchecks that should not be run inline with the healthcheck request.
For example, verifying that downstream services are up should be done
as part of a separate continuous process which can be interrogated during
a healthcheck request to get the most recent status of that check.